Unified flow extraction for ASN Sankey and tables
